Stand under platform, hot-cold game rewards, in practice doesn't work
The issue is, at the start the network has no idea about any good states, but soon it figures out you can run into wall and get a consistent reward. A reward is better than no reward, so it hangs around wall. Why is it unable to wander into better states with epsilon-greedy? Because even if it does, the few out of X replay-buffered states that wandered in better direction will get the Qvalue estimation fixed, but other 100 will bomb the estimation with old value, so the prediction won't budge. Solution: don't use this reward system?...