removed useless lrate for qlearning
it's not needed when there already is learning late for gradient descent!!! Makes so much sense. I also changed normalisation for platforms, because i thought the platforms end at "40", but actually it's only the long platforms and they extend past screen, it's better to make it "35". However, in this commit there is no reward for approaching target, only for being near, so again it learns to run past target and hugs wall, at least this wasn't unexpected i guess... but there needs to be a reward that will discourage this