Major success, harold learns to stand under platform :D
So I thought, when the platforms don't change, it's not really any easier for agent to learn the task, because it's almost impossible to generate diverse training data. When platforms are randomised the agent doesn't focus on one platform, right, but it's not the point, the point is to learn to interpret the platform coordinates. So randomised games provide very broad data and the learning works without tricks. Issue: the rewards don't get higher for standing closer to middle, hence the agent learned to stay on the right edge, which still counts, fair. Worse issue: it really doesn't want to learn that no input is good, because exploration isn't good enough...